[D83T-86] 내 눈바디 남기기 수정 API구현 + 내 눈바디 남기기 등록 API 파일이 없는 경우에도 요청 받을 수 있도록 수정

+++ src/main/java/d83t/bpmbackend/domain/aggregate/community/service/BodyShapeServiceImpl.java
@@ -64,12 +64,13 @@ public class BodyShapeServiceImpl implements BodyShapeService {
     @Override
     public BodyShapeResponse createBoastArticle(User user, List<MultipartFile> files, BodyShapeRequest bodyShapeRequest) {
         //file은 최대 5개만 들어올 수 있다.
-        if (files.size() > 5) {
+        if (files != null && files.size() > 5) {
             throw new CustomException(Error.FILE_SIZE_MAX);
         }
 
-        User findUser = userRepository.findByKakaoId(user.getKakaoId()).orElseThrow(() -> {throw new CustomException(Error.NOT_FOUND_USER_ID);});
-        List<String> filePaths = new ArrayList<>();
+        User findUser = userRepository.findByKakaoId(user.getKakaoId()).orElseThrow(() -> {
+            throw new CustomException(Error.NOT_FOUND_USER_ID);
+        });
 
         Profile profile = findUser.getProfile();
         BodyShape bodyshape = BodyShape.builder()
@@ -77,24 +78,27 @@ public class BodyShapeServiceImpl implements BodyShapeService {
                 .content(bodyShapeRequest.getContent())
                 .build();
 
-        for (MultipartFile file : files) {
-            String newName = FileUtils.createNewFileName(file.getOriginalFilename());
-            String filePath = fileDir + newName;
-            bodyshape.addBodyShapeImage(BodyShapeImage.builder()
-                    .originFileName(newName)
-                    .storagePathName(filePath)
-                    .bodyShape(bodyshape)
-                    .build());
-            filePaths.add(filePath);
-            if (env.equals("prod")) {
-                uploaderService.putS3(file, bodyShapePath, newName);
-            } else if (env.equals("local")) {
-                try {
-                    File localFile = new File(filePath);
-                    file.transferTo(localFile);
-                    FileUtils.removeNewFile(localFile);
-                } catch (IOException e) {
-                    e.printStackTrace();
+        List<String> filePaths = new ArrayList<>();
+        if(files != null && files.size() != 0) {
+            for (MultipartFile file : files) {
+                String newName = FileUtils.createNewFileName(file.getOriginalFilename());
+                String filePath = fileDir + newName;
+                bodyshape.addBodyShapeImage(BodyShapeImage.builder()
+                        .originFileName(newName)
+                        .storagePathName(filePath)
+                        .bodyShape(bodyshape)
+                        .build());
+                filePaths.add(filePath);
+                if (env.equals("prod")) {
+                    uploaderService.putS3(file, bodyShapePath, newName);
+                } else if (env.equals("local")) {
+                    try {
+                        File localFile = new File(filePath);
+                        file.transferTo(localFile);
+                        FileUtils.removeNewFile(localFile);
+                    } catch (IOException e) {
+                        e.printStackTrace();
+                    }
                 }
             }
         }
@@ -147,13 +151,13 @@ public class BodyShapeServiceImpl implements BodyShapeService {
     //TODO: 향후 public, private 구분이 되는 기능이 추가될 시 Response를 바꿔야함.
     @Override
     public BodyShapeResponse getBodyShape(User user, Long bodyShapeId) {
-        BodyShape bodyShape = bodyShapeRepository.findById(bodyShapeId).orElseThrow(()->{
+        BodyShape bodyShape = bodyShapeRepository.findById(bodyShapeId).orElseThrow(() -> {
             throw new CustomException(Error.NOT_FOUND_BODY_SHAPE);
         });
         List<String> filePaths = new ArrayList<>();
         Profile author = bodyShape.getAuthor();
         List<BodyShapeImage> images = bodyShape.getImages();
-        for(BodyShapeImage image : images){
+        for (BodyShapeImage image : images) {
             filePaths.add(image.getStoragePathName());
         }
         return BodyShapeResponse.builder()
@@ -169,11 +173,79 @@ public class BodyShapeServiceImpl implements BodyShapeService {
                 .build();
     }
 
+    //TODO: 향후 public에 대한 대처가 필요.
+    @Override
+    public BodyShapeResponse updateBodyShape(User user, List<MultipartFile> files, BodyShapeRequest bodyShapeRequest, Long bodyShapeId) {
+        if (files != null && files.size() > 5) {
+            throw new CustomException(Error.FILE_SIZE_MAX);
+        }
+        User findUser = userRepository.findByKakaoId(user.getKakaoId()).orElseThrow(() -> {
+            throw new CustomException(Error.NOT_FOUND_USER_ID);
+        });
+
+        BodyShape bodyShape = bodyShapeRepository.findById(bodyShapeId).orElseThrow(() -> {
+            throw new CustomException(Error.NOT_FOUND_BODY_SHAPE);
+        });
+
+        if (bodyShapeRequest.getContent() != null) {
+            bodyShape.changeContent(bodyShapeRequest.getContent());
+        }
+
+        List<BodyShapeImage> bodyShapeImages = bodyShape.getImages();
+        List<String> filePaths = bodyShapeImages.stream()
+                .map(BodyShapeImage::getStoragePathName)
+                .collect(Collectors.toList());
+
+        //파일 수정
+        if (files != null && files.size() != 0) {
+            filePaths = new ArrayList<String>();
+            List<BodyShapeImage> updateBodyShapeImages = new ArrayList<>();
+            for (MultipartFile file : files) {
+                String newName = FileUtils.createNewFileName(file.getOriginalFilename());
+                String filePath = fileDir + newName;
+
+                updateBodyShapeImages.add(BodyShapeImage.builder()
+                        .originFileName(newName)
+                        .storagePathName(filePath)
+                        .bodyShape(bodyShape)
+                        .build());
+
+                filePaths.add(filePath);
+                if (env.equals("prod")) {
+                    uploaderService.putS3(file, bodyShapePath, newName);
+                } else if (env.equals("local")) {
+                    try {
+                        File localFile = new File(filePath);
+                        file.transferTo(localFile);
+                        FileUtils.removeNewFile(localFile);
+                    } catch (IOException e) {
+                        e.printStackTrace();
+                    }
+                }
+            }
+            bodyShape.changeImage(updateBodyShapeImages);
+        }
+
+        Profile profile = findUser.getProfile();
+
+        return BodyShapeResponse.builder()
+                .id(bodyShape.getId())
+                .createdAt(bodyShape.getCreatedDate())
+                .updatedAt(bodyShape.getModifiedDate())
+                .author(BodyShapeResponse.Author.builder()
+                        .nickname(profile.getNickName())
+                        .profilePath(profile.getStoragePathName())
+                        .build())
+                .filesPath(filePaths)
+                .content(bodyShape.getContent())
+                .build();
+    }
+
     //TODO: 마찬가지로 향후 public이 생길시 작성 유저인지 판단하는 로직 추가
     @Override
     public void deleteBodyShape(User user, Long bodyShapeId) {
         Optional<BodyShape> findBodyShape = bodyShapeRepository.findById(bodyShapeId);
-        if(findBodyShape.isEmpty()){
+        if (findBodyShape.isEmpty()) {
             throw new CustomException(Error.NOT_FOUND_BODY_SHAPE);
         }
         BodyShape bodyShape = findBodyShape.get();
